@@ -21,67 +21,73 @@
     "*.{css,vue}": "stylelint"
   },
   "devDependencies": {
-    "@nuxt/ui": "^2.17.0",
-    "@nuxtjs/i18n": "^9.3.0",
-    "@nuxtjs/stylelint-module": "^5.2.0",
-    "@tailwindcss/typography": "^0.5.12",
-    "@types/jest": "^29.5.12",
-    "@typescript-eslint/eslint-plugin": "^5.62.0",
-    "@typescript-eslint/parser": "^5.62.0",
-    "@vitest/coverage-istanbul": "^1.6.0",
-    "@vitest/coverage-v8": "^1.6.0",
-    "@vueuse/nuxt": "^10.9.0",
-    "eslint": "^8.57.0",
+    "@eslint/eslintrc": "^3.3.1",
+    "@eslint/js": "^9.25.1",
+    "@nuxt/test-utils": "^3.17.2",
+    "@nuxt/ui": "^2.21.0",
+    "@nuxtjs/i18n": "^9.5.3",
+    "@nuxtjs/stylelint-module": "^5.2.1",
+    "@nuxtjs/tailwindcss": "6.13.1",
+    "@tailwindcss/typography": "^0.5.16",
+    "@types/node": "^22.15.2",
+    "@typescript-eslint/eslint-plugin": "^8.31.0",
+    "@typescript-eslint/parser": "^8.31.0",
+    "@vitest/coverage-v8": "^3.1.2",
+    "@vitest/ui": "^3.1.2",
+    "@vue/test-utils": "^2.4.6",
+    "@vueuse/nuxt": "^10.11.1",
+    "eslint": "8.57.0",
     "eslint-config-prettier": "^8.10.0",
     "eslint-plugin-nuxt": "^4.0.0",
     "eslint-plugin-prettier": "^4.2.1",
-    "eslint-plugin-vue": "^9.23.0",
+    "eslint-plugin-vue": "9.23.0",
     "eslint-plugin-vue-pug": "^0.5.5",
+    "globals": "^16.0.0",
     "husky": "^8.0.3",
-    "jsdom": "^21.1.2",
+    "jsdom": "^26.1.0",
     "lint-staged": "^13.3.0",
-    "nuxt": "^3.11.1",
+    "nuxt": "^3.16.2",
     "nuxt-icon": "^0.2.11",
-    "nuxt-tiptap-editor": "^1.1.0",
-    "postcss-html": "^1.6.0",
+    "nuxt-tiptap-editor": "^1.2.1",
+    "postcss-html": "^1.8.0",
     "postcss-import": "^15.1.0",
     "prettier": "^2.8.8",
-    "pug": "^3.0.2",
+    "pug": "^3.0.3",
     "pug-plain-loader": "^1.1.0",
-    "stylelint": "^16.6.1",
-    "stylelint-config-recommended-vue": "^1.5.0",
+    "stylelint": "^16.19.1",
+    "stylelint-config-recommended-vue": "^1.6.0",
     "stylelint-config-standard": "^36.0.1",
+    "typescript": "^5.8.3",
     "unplugin-auto-import": "^0.16.7",
     "vite-tsconfig-paths": "^4.3.2",
-    "vitest": "^1.6.0",
-    "vue-tsc": "^1.8.27"
+    "vitest": "^3.1.2",
+    "vue-eslint-parser": "^10.1.3",
+    "vue-tsc": "^2.2.10"
   },
   "dependencies": {
-    "@headlessui/vue": "^1.7.19",
+    "@headlessui/vue": "^1.7.23",
     "@mdi/font": "^7.4.47",
     "@pinia/nuxt": "^0.4.11",
     "@types/d3": "^7.4.3",
-    "@unhead/shared": "^1.8.20",
+    "@unhead/shared": "^1.11.20",
     "@vuepic/vue-datepicker": "^5.4.0",
-    "@vueuse/core": "^10.9.0",
-    "axios": "^1.6.8",
+    "@vueuse/core": "^10.11.1",
+    "axios": "^1.9.0",
     "d3": "^7.9.0",
     "date-fns": "^3.6.0",
-    "install": "^0.13.0",
     "jsbarcode": "^3.11.6",
-    "keycloak-js": "^24.0.1",
-    "pinia": "^2.1.7",
-    "sass": "^1.72.0",
-    "typescript": "^5.4.3",
+    "keycloak-js": "^24.0.5",
+    "pinia": "^2.3.1",
+    "sass": "^1.87.0",
     "v-calendar": "^3.1.2",
     "v3-infinite-loading": "^1.3.2",
     "vue-composable": "1.0.0-beta.24",
-    "vue-currency-input": "^3.1.0",
+    "vue-currency-input": "^3.2.1",
     "vue-qr": "^4.0.9",
-    "vue-router": "^4.3.0",
+    "vue-router": "^4.5.1",
     "vuedraggable": "^4.1.0"
   },
   "volta": {
-    "node": "18.18.2"
+    "node": "20.9.0"
   }
 }
